pyproject.toml
[project]
name = "pineventory"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "discord.py",
    "python-dotenv",
    "pydantic",
    "asyncpg",
    "gspread",
    "oauth2client",
    "google-api-python-client",
    "google-auth-oauthlib",
    "pytest-asyncio",
    "aiohttp"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
timeout = 10
